archinstall/requirements.txt
python
python-rich
python-inquirer
